% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_masic_labeled.R
\name{as_masic_labeled}
\alias{as_masic_labeled}
\title{Make a MASIC data object for labeled data}
\usage{
as_masic_labeled(masic)
}
\arguments{
\item{masic}{(data.frame) The data.frame to be converted into a labeled masic data object. Required.}
}
\value{
(masic_data_labeled) A labeled masic data object.
}
\description{
Make a MASIC data object for labeled data
}
