% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MSGF.R
\name{ascore_filter}
\alias{ascore_filter}
\title{Filter msnid by an ascore}
\usage{
ascore_filter(msnid, ascore_file, ascore_threshold = 17)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}

\item{ascore_file}{(data.frame) Add an AScore file to filter phosphoproteomics data by uploading
this data.frame. Default is NULL.}

\item{ascore_threshold}{(numeric) If phosphoproteomics, filter by an AScore threshold.
Default is 17, see \href{https://www.nature.com/articles/nbt1240}{here}}
}
\value{
(MSnID object) Filtered MSGF output
}
\description{
Requires an extra ascore_file that can be pulled with get_Ascore. Any NA AScores
are maintained.
}
