% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_e_data}
\alias{create_e_data}
\title{Create an e_data (biomolecule expression or crosstab) from masic data, msnid, and f_data}
\usage{
create_e_data(masic, msnid, f_data)
}
\arguments{
\item{masic}{A masic_data object that has been filtered by intereference_filter. Required.}

\item{msnid}{An msnid object that has been fdr and decoy filtered. Required.}

\item{f_data}{An f_data object from create_f_data. Required.}
}
\value{
A pmartR e_data (biomolecule expression) object
}
\description{
Create an e_data (biomolecule expression or crosstab) from masic data, msnid, and f_data
}
