% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_e_objects_labeled}
\alias{create_e_objects_labeled}
\title{Create an e_data (biomolecule expression or crosstab) and e_meta (biomolecule data) from labeled masic data, msnid, f_data, and plex information}
\usage{
create_e_objects_labeled(masic, msnid, f_data, plex_data)
}
\arguments{
\item{masic}{A masic_data_labeled object that has been filtered by intereference_filter. Required.}

\item{msnid}{An msnid object that has been fdr and decoy filtered. Required.}

\item{f_data}{An f_data object from create_f_data. Contains at a minimum the
SampleNames, PlexNames, and IonChannelNames. Required.}

\item{plex_data}{Contains at a minimum the Dataset and PlexNames information. Both
f_data and plex_data are mapped to their plex and ion channels for merging. Required.}
}
\value{
A list with a pmartR e_data (biomolecule expression) object and a pmartR e_meta (biomolecule data) object
}
\description{
Create an e_data (biomolecule expression or crosstab) and e_meta (biomolecule data) from labeled masic data, msnid, f_data, and plex information
}
