% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_e_objects_unlabeled}
\alias{create_e_objects_unlabeled}
\title{Create an e_data (biomolecule expression or crosstab) and e_meta (biomolecule data) from unlabeled masic data, msnid, and metadata}
\usage{
create_e_objects_unlabeled(masic, msnid, metadata)
}
\arguments{
\item{masic}{A masic_data_unlabeled object. Required.}

\item{msnid}{An msnid object that has been fdr and decoy filtered. Required.}

\item{metadata}{A data frame. SampleNames and FileNames are required,
where any duplicated SampleNames are summed.}
}
\value{
A list with a pmartR e_data (biomolecule expression) object and a pmartR e_meta (biomolecule data) object
}
\description{
Create an e_data (biomolecule expression or crosstab) and e_meta (biomolecule data) from unlabeled masic data, msnid, and metadata
}
