% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_f_data}
\alias{create_f_data}
\title{Create an f_data (sample information object) from masic data and a metadata file}
\usage{
create_f_data(masic, metadata)
}
\arguments{
\item{masic}{A masic_data object that has been filtered by intereference_filter. Required.}

\item{metadata}{A data frame with the PlexNames, IonChannelNames, SampleName,
and any other metadata to end up in the f_data file. Know that PlexNames and
IonChannelNames are used to find the data in the masic table, and Sample Name
is what the column names will be in the final table. Required.}
}
\value{
A pmartR f_data (sample information) object
}
\description{
Create an f_data (sample information object) from masic data and a metadata file
}
