% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MSGF.R
\name{decoy_filter}
\alias{decoy_filter}
\title{Remove decoys}
\usage{
decoy_filter(msnid)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}
}
\value{
(MSnID object) Filtered MSGF output
}
\description{
Filtering decoys from an msnid object AFTER the FDR filter has been applied.
}
