% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MSGF.R
\name{fdr_filter}
\alias{fdr_filter}
\title{Filter msnid object to a specific FDR}
\usage{
fdr_filter(
  msnid,
  level = "peptide",
  fdr.max = 0.01,
  n.iter.grid = 500,
  n.iter.nm = 100
)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}

\item{level}{(character) Level at which to perform FDR filter. Options are peptide and accession (protein). Default is peptide.}

\item{fdr.max}{(numeric) Maximum acceptable FDR rate. Default is 0.01.}

\item{n.iter.grid}{(numeric) Bumber of grid-distributed evaluation points. Default 500.}

\item{n.iter.nm}{(numeric) Number of iterations for Nelder-Mead optimization algorithm. Default is 100.}
}
\value{
(MSnID object) Filtered MSGF output
}
\description{
Filtering MSGF data. In this implementation the peptide level filter optimizes both ppm and
PepQValue thresholds to achieve maximum number of peptide identifications within
the given FDR constraints. https://github.com/vladpetyuk/PlexedPiper/blob/master/R/filter_msgf_data.R
}
