% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MASIC.R
\name{interference_filter}
\alias{interference_filter}
\title{Filter MASIC data by an interference score and remove all missing measurements}
\usage{
interference_filter(
  masic,
  interference_score_threshold = 0.9,
  remove_all_missing = TRUE
)
}
\arguments{
\item{masic}{An object of the masic data class. Required.}

\item{interference_score_threshold}{A numeric between 0-1 to view the maximum interference.
The higher the number, the cleaner parent ion at MS1 level. Default is 0.5.}

\item{remove_all_missing}{A logical to indicate whether a dataset and scan number
where there is no intensity/abundance at any ion intensity should be removed.}
}
\value{
A filtered object of the masic data class
}
\description{
Filter MASIC data by an interference score and remove all missing measurements
}
