% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MASIC_labeled.R
\name{interference_filter_labeled}
\alias{interference_filter_labeled}
\title{Filter MASIC data by an interference score and remove all missing measurements}
\usage{
interference_filter_labeled(masic, interference_score_threshold = 0.5)
}
\arguments{
\item{masic}{An object of the masic_data_isobaric class. Required.}

\item{interference_score_threshold}{A numeric between 0-1 to view the maximum interference.
The higher the number, the cleaner parent ion at MS1 level. Default is 0.5.}
}
\value{
A filtered object of the masic data class
}
\description{
Filter MASIC data by an interference score and remove all missing measurements
}
