% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_labeled_pipeline.R
\name{run_labeled_pipeline}
\alias{run_labeled_pipeline}
\title{Main labeled pipeline script}
\usage{
run_labeled_pipeline(
  msnid,
  masic,
  metadata,
  plex_data,
  outpath = "~/Downloads/",
  interference_score_threshold = 0.5,
  ascore_file = NULL,
  ascore_threshold = 17
)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}

\item{masic}{An object of the masic data class. Required.}

\item{metadata}{A data frame with the IonChannelNames, PlexNames, and SampleNames,
and any other metadata to end up in the f_data file. SampleNames will
replace the IonChannelNames and PlexNames combinations.}

\item{plex_data}{Contains at a minimum the Dataset and PlexNames information.
Both f_data and plex_data are mapped to their plex and ion channels for merging. Required.}

\item{outpath}{The folder to write the e_data, e_meta, and f_data files to. Required.}

\item{interference_score_threshold}{A numeric between 0-1 to view the maximum interference.
The higher the number, the cleaner parent ion at MS1 level. Default is 0.5.}

\item{ascore_file}{(data.frame) Add an AScore file to filter phosphoproteomics data by uploading
this data.frame. Default is NULL.}

\item{ascore_threshold}{(numeric) If phosphoproteomics, filter by an AScore threshold.
Default is 17, see \href{https://www.nature.com/articles/nbt1240}{here}}
}
\description{
Main labeled pipeline script
}
