% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_labeled_pipeline.R
\name{tmt_pipeline}
\alias{tmt_pipeline}
\title{Main pipeline script to run the TMT Pipeline}
\usage{
tmt_pipeline(
  msnid,
  masic,
  metadata,
  plex_data,
  outpath = "~/Downloads/",
  interference_score_threshold = 0.5,
  a_score_threshold = NULL
)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}

\item{masic}{An object of the masic data class. Required.}

\item{metadata}{A data frame with the IonChannelNames, PlexNames, and SampleNames,
and any other metadata to end up in the f_data file. SampleNames will
replace the IonChannelNames and PlexNames combinations.}

\item{plex_data}{Contains at a minimum the Dataset and PlexNames information.
Both f_data and plex_data are mapped to their plex and ion channels for merging. Required.}

\item{outpath}{The folder to write the e_data, e_meta, and f_data files to. Required.}

\item{interference_score_threshold}{A numeric between 0-1 to view the maximum interference.
The higher the number, the cleaner parent ion at MS1 level. Default is 0.5.}

\item{ascore_threshold}{If phosphoproteomics, filter by an A score threshold.
Default is NULL.}
}
\description{
Main pipeline script to run the TMT Pipeline
}
