% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_unlabeled_pipeline.R
\name{unlabeled_pipeline}
\alias{unlabeled_pipeline}
\title{Main unlabeled pipeline script}
\usage{
unlabeled_pipeline(
  msnid,
  folder_path,
  metadata,
  outpath = "~/Downloads/",
  interference_score_threshold = 0.5
)
}
\arguments{
\item{msnid}{(MSnID object) collated MSGF output. Required.}

\item{folder_path}{Path to the SICstats txt files to generate the masic dataset. Required.}

\item{metadata}{A data frame with the FileNames and the SampleNames. Required.}

\item{outpath}{The folder to write the e_data, e_meta, and f_data files to. Required.}

\item{interference_score_threshold}{A numeric between 0-1 to view the maximum interference.
The higher the number, the cleaner parent ion at MS1 level. Default is 0.5.}
}
\description{
Main unlabeled pipeline script
}
