% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_masic_labeled.R
\name{as_masic_labeled}
\alias{as_masic_labeled}
\title{Make a MASIC data object for labeled data}
\usage{
as_masic_labeled(masic)
}
\arguments{
\item{masic}{The data to make a labeled masic data object. Required.}
}
\description{
Make a MASIC data object for labeled data
}
