% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_f_data}
\alias{create_f_data}
\title{Create an f_data (sample information object) from masic data and a metadata file}
\usage{
create_f_data(masic, metadata)
}
\arguments{
\item{masic}{A masic_data_unlabeled or a masic_data_labeled object. Required.}

\item{metadata}{A data frame. For labeled proteomics, the minimum columns:
IonChannelNames, PlexNames, and SampleNames, where SampleNames
will replace the IonChannelNames and PlexNames combinations. For unlabeled
proteomics, SampleNames and FileNames are required, where any duplicated SampleNames
are summed. Any additional columns in metadata will be returned in the
f_data file. Required.}
}
\value{
A pmartR f_data (sample information) object
}
\description{
Create an f_data (sample information object) from masic data and a metadata file
}
