% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pmartR_objects.R
\name{create_f_data}
\alias{create_f_data}
\title{Create an f_data (sample information object) from masic data and a metadata file}
\usage{
create_f_data(masic, metadata)
}
\arguments{
\item{masic}{A masic_data object that has been filtered by intereference_filter. Required.}

\item{metadata}{A data frame with the IonChannelNames, PlexNames, and SampleNames,
and any other metadata to end up in the f_data file. SampleNames
will be replace the IonChannelNames and PlexNames combinations.}
}
\value{
A pmartR f_data (sample information) object
}
\description{
Create an f_data (sample information object) from masic data and a metadata file
}
